logic moved from MovieTile into MovieList
